renamed src/utils/catchError.ts
